+++ C++/Gatto/gatto.txt
